feat: Fetch and use YouTube visitor data
This commit adds functionality to fetch and utilize YouTube's visitor data. This data is retrieved from the `https://www.youtube.com/sw.js_data` endpoint and stored in `ytContext.client.visitorData`. This data is then used to improve the retrieval of video streams.

Additionally, the commit updates the client used for retrieving streams to IOS.